+++ src/main/java/de/stklcode/jvault/connector/model/AuthBackend.java
@@ -24,8 +24,6 @@ package de.stklcode.jvault.connector.model;
  */
 public enum AuthBackend {
     TOKEN("token"),
-    @Deprecated(since = "1.1.3", forRemoval = true)
-    APPID("app-id"),
     APPROLE("approle"),
     USERPASS("userpass"),
     GITHUB("github"),   // Not supported yet.